# docs/specs/web/openapi.yaml
openapi: 3.1.0
info:
  title: MTG Deck Tools API
  description: HTTP surface for the MTG Deck Tools engine. Validation and deck rules
    run server-side; clients send DeckCriteria and receive .deck.json-shaped documents.
  version: 0.1.0
paths:
...
  /api/v1/decks:
    get:
      tags:
      - library
      summary: List Decks
      operationId: list_decks_api_v1_decks_get
      parameters:
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search name, commander, or themes
          title: Q
        description: Search name, commander, or themes
      - name: sort
        in: query
        required: false
        schema:
          type: string
          description: saved_at, name, or commander
          default: saved_at
          title: Sort
        description: saved_at, name, or commander
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: db
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: SQLite database path
          title: Db
        description: SQLite database path
      - name: decks
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Saved deck library path
          title: Decks
        description: Saved deck library path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeckLibraryEntry'
                title: Response List Decks Api V1 Decks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decks/{deck_id}:
    get:
      tags:
      - library
      summary: Get Deck
      operationId: get_deck_api_v1_decks__deck_id__get
      parameters:
      - name: deck_id
        in: path
        required: true
        schema:
          type: string
          title: Deck Id
      - name: db
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: SQLite database path
          title: Db
        description: SQLite database path
      - name: decks
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Saved deck library path
          title: Decks
        description: Saved deck library path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckLibraryDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - library
      summary: Patch Deck
      operationId: patch_deck_api_v1_decks__deck_id__patch
      parameters:
      - name: deck_id
        in: path
        required: true
        schema:
          type: string
          title: Deck Id
      - name: db
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: SQLite database path
          title: Db
        description: SQLite database path
      - name: decks
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Saved deck library path
          title: Decks
        description: Saved deck library path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchDeckRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckLibraryDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - library
      summary: Delete Deck
      operationId: delete_deck_api_v1_decks__deck_id__delete
      parameters:
      - name: deck_id
        in: path
        required: true
        schema:
          type: string
          title: Deck Id
      - name: db
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: SQLite database path
          title: Db
        description: SQLite database path
      - name: decks
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Saved deck library path
          title: Decks
        description: Saved deck library path
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decks/{deck_id}/refill-slot:
    post:
      tags:
      - library
      summary: Refill Deck Slot
      operationId: refill_deck_slot_api_v1_decks__deck_id__refill_slot_post
      parameters:
      - name: deck_id
        in: path
        required: true
        schema:
          type: string
          title: Deck Id
      - name: db
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: SQLite database path
          title: Db
        description: SQLite database path
      - name: decks
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Saved deck library path
          title: Decks
        description: Saved deck library path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefillSlotRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckLibraryDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decks/{deck_id}/swap:
    post:
      tags:
      - library
      summary: Swap Deck Cards
      operationId: swap_deck_cards_api_v1_decks__deck_id__swap_post
      parameters:
      - name: deck_id
        in: path
        required: true
        schema:
          type: string
          title: Deck Id
      - name: db
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: SQLite database path
          title: Db
        description: SQLite database path
      - name: decks
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Saved deck library path
          title: Decks
        description: Saved deck library path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwapCardsRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwapCardsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DeckLibraryDetailResponse:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        saved_at:
          type: string
          title: Saved At
        deck:
          additionalProperties: true
          type: object
          title: Deck
      type: object
      required:
      - id
      - name
      - saved_at
      - deck
      title: DeckLibraryDetailResponse
    DeckLibraryEntry:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        saved_at:
          type: string
          title: Saved At
        commander_names:
          items:
            type: string
          type: array
          title: Commander Names
        commander_image_uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Commander Image Uri
        colors:
          items:
            type: string
          type: array
          title: Colors
        themes:
          items:
            type: string
          type: array
          title: Themes
        estimated_price_usd:
          anyOf:
          - type: number
          - type: 'null'
          title: Estimated Price Usd
      type: object
      required:
      - id
      - name
      - saved_at
      title: DeckLibraryEntry
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PatchDeckRequest:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        deck:
          anyOf:
          - type: object
            additionalProperties: true
          - type: 'null'
          title: Deck
      type: object
      title: PatchDeckRequest
    RefillSlotRequest:
      properties:
        slot:
          type: string
          title: Slot
        seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seed
      type: object
      required:
      - slot
      title: RefillSlotRequest
    SwapCardsRequest:
      properties:
        oracle_ids:
          items:
            type: string
          type: array
          title: Oracle Ids
        seed:
          anyOf:
          - type: integer
          - type: 'null'
          title: Seed
      type: object
      required:
      - oracle_ids
      title: SwapCardsRequest
    SwapRecordResponse:
      properties:
        slot:
          type: string
          title: Slot
        from_oracle_id:
          type: string
          title: From Oracle Id
        from_name:
          type: string
          title: From Name
        to_oracle_id:
          type: string
          title: To Oracle Id
        to_name:
          type: string
          title: To Name
      type: object
      required:
      - slot
      - from_oracle_id
      - from_name
      - to_oracle_id
      - to_name
      title: SwapRecordResponse
    SwapCardsResponse:
      properties:
        id:
          type: string
          title: Id
        deck:
          type: object
          additionalProperties: true
          title: Deck
        swaps:
          items:
            $ref: '#/components/schemas/SwapRecordResponse'
          type: array
          title: Swaps
      type: object
      required:
      - id
      - deck
      title: SwapCardsResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...